refactor(apps): create dedicated storybook app
