+++ apps/stories/.storybook/waystone.decorator.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource theme-ui */
 import { useEffect } from 'react';
 import { useColorMode } from 'theme-ui';
-import { Waystone } from '../src';
+import { Waystone } from '@waystone/components/src';
 import { useDarkMode } from 'storybook-dark-mode';
 
 const ThemeChanger = () => {
