+++ apps/stories/public/example-image.png
