+++ apps/stories/src/design-system/colors.stories.tsx
@@ -1,8 +1,8 @@
 import React, { FC, PropsWithChildren } from 'react';
 import { Meta, Story } from '@storybook/react';
-import { WaystoneBaseTheme } from '@waystone/core';
+import { WaystoneBaseTheme } from '@waystone/core/src';
 import { ColorMode, ColorModesScale } from 'theme-ui';
-import { Text } from '..';
+import { Text } from '@waystone/components/src';
 
 const Color: FC<PropsWithChildren<{ color: string }>> = ({
   color,
