+++ apps/stories/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "extends": "@waystone/tsconfig/react-library.json",
+  "include": ["src"],
+  "exclude": [
+    "dist",
+    "build",
+    "node_modules",
+    "storybook-static",
+    ".storybook",
+    "__snapshots__",
+    "__assets__",
+    "__tests__",
+    "__fixtures__",
+    "**/*.stories.tsx",
+    "**/*.{spec,test,internal}.tsx"
+  ],
+  "compilerOptions": {
+    "outDir": "dist"
+  }
+}
