+++ packages/waystone-components/package.json
@@ -17,34 +17,20 @@
   ],
   "scripts": {
     "build": "tsup src/index.tsx --format esm,cjs --dts --external react",
-    "build:storybook": "build-storybook",
     "dev": "tsup src/index.tsx --format esm,cjs --watch --dts --external react",
-    "dev:storybook": "start-storybook -p 6006",
     "lint": "eslint src --fix",
     "lint:release": "eslint src",
-    "clean": "rimraf -rf .turbo && rimraf -rf node_modules && rimraf -rf dist && rimraf -rf storybook-static && rimraf -rf coverage",
+    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist && rm -rf storybook-static && rm -rf coverage",
     "test": "jest --coverage",
     "test:release": "jest --coverage",
     "test:update": "jest --coverage -u",
-    "test:watch": "jest --coverage --watch -u",
-    "storybook:upgrade": "npx sb@latest upgrade -s"
+    "test:watch": "jest --coverage --watch -u"
   },
   "devDependencies": {
-    "@storybook/addon-a11y": "^6.5.16",
-    "@storybook/addon-actions": "^6.5.16",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-interactions": "^6.5.16",
-    "@storybook/addon-links": "^6.5.16",
-    "@storybook/builder-webpack5": "^6.5.16",
-    "@storybook/jest": "^0.0.10",
-    "@storybook/manager-webpack5": "^6.5.16",
-    "@storybook/react": "^6.5.16",
-    "@storybook/testing-library": "^0.0.13",
     "@types/react-syntax-highlighter": "^15.5.5",
     "@waystone/tsconfig": "*",
     "eslint-preset-waystone": "*",
-    "identity-obj-proxy": "^3.0.0",
-    "storybook-dark-mode": "^2.1.1"
+    "identity-obj-proxy": "^3.0.0"
   },
   "dependencies": {
     "@emotion/react": "^11.10.4",
