+++ packages/waystone-utils/package.json
@@ -22,7 +22,7 @@
     "dev:storybook": "tsup src/index.tsx --format esm,cjs --watch --dts --external react",
     "lint": "eslint src --fix",
     "lint:release": "eslint src",
-    "clean": "rimraf -rf .turbo && rimraf -rf node_modules && rimraf -rf dist"
+    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
   },
   "dependencies": {
     "react": "^18.0.0",
